Move postprocessing code into postprocessing directory (IDE refactor, no manual changes)
