+++ metaspace/engine/sm/engine/postprocessing/ion_thumbnail.py
